backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyjwt
passlib
bcrypt
mistralai
elevenlabs
moviepy
python-dotenv
python-multipart
